investor-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
coverage[toml]==7.10.1
    # via pytest-cov
dnspython==2.7.0
    # via pymongo
factory-boy==3.3.3
    # via -r requirements.in
faker==37.4.2
    # via
    #   -r requirements.in
    #   factory-boy
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
iniconfig==2.1.0
    # via pytest
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pymongo[srv]==4.10.1
    # via -r requirements.in
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
pytest-asyncio==1.1.0
    # via -r requirements.in
pytest-cov==6.2.1
    # via -r requirements.in
pytest-mock==3.14.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
redis==6.2.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via faker
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
